+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - images: fix format of dists hash and centos/rhel sources loop for appstream
 - images: fix group/extra packages installation order
 
+### Changed
+- images: bootstrap RPM images with generated dnf/yum repos files
+
 ## [0.20220125] - 2022-01-25
 
 ### Added